+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added a warning about old projects not being available
 - quick font reexport: easily export subsequent changes
 - require space in the character set
+- added visual hint for character sizing
 
 ## 2.2.0